did some stuff with the calories page. calories are stored in shared preferences. I suck at UI formatting

+++ app/src/main/java/com/app/agile_overlords/moveandgroove/Models/UserModel.java
@@ -7,7 +7,7 @@ public class UserModel {
 
     public UserModel(){}
     //Todo: This was causing a conflict. Weight is a float here, but Integer else where
-    public UserModel(String sex, int age, float weight){
+    public UserModel(String sex, int age, double weight){
         this.sex = sex;
         this.age = age;
         //this.weight = weight;
